requirements.txt
praw
pymongo
python-dotenv
selenium
streamlit
appium-python-client
google-generativeai